requirements.txt
genie-tts
python-dotenv
